rename and move files around to diff packages
