+++ app/src/main/java/com/github/odaridavid/weatherapp/designsystem/PermissionComponent.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.weatherapp.ui
+package com.github.odaridavid.weatherapp.designsystem
 
 import android.Manifest
 import android.app.Activity
