+++ app/src/main/java/com/github/odaridavid/weatherapp/designsystem/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.weatherapp.ui.theme
+package com.github.odaridavid.weatherapp.designsystem.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
