+++ app/src/main/java/com/github/odaridavid/weatherapp/designsystem/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.weatherapp.ui.theme
+package com.github.odaridavid.weatherapp.designsystem.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
