+++ app/src/main/java/com/github/odaridavid/weatherapp/ui/AppNavGraph.kt
@@ -2,13 +2,13 @@ package com.github.odaridavid.weatherapp.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.github.odaridavid.weatherapp.ui.home.HomeScreen
 import com.github.odaridavid.weatherapp.ui.home.HomeScreenIntent
 import com.github.odaridavid.weatherapp.ui.home.HomeViewModel
-import com.github.odaridavid.weatherapp.ui.home.HomeScreenViewState
 import com.github.odaridavid.weatherapp.ui.settings.SettingsScreen
 import com.github.odaridavid.weatherapp.ui.settings.SettingsScreenIntent
 import com.github.odaridavid.weatherapp.ui.settings.SettingsScreenViewState
@@ -16,12 +16,11 @@ import com.github.odaridavid.weatherapp.ui.settings.SettingsViewModel
 
 @Composable
 fun WeatherAppScreensConfig(
-    navController: NavHostController,
-    homeViewModel: HomeViewModel,
-    settingsViewModel: SettingsViewModel
+    navController: NavHostController
 ) {
     NavHost(navController = navController, startDestination = Destinations.HOME.route) {
         composable(Destinations.HOME.route) {
+            val homeViewModel = hiltViewModel<HomeViewModel>()
             val state = homeViewModel
                 .state
                 .collectAsState()
@@ -39,6 +38,7 @@ fun WeatherAppScreensConfig(
             )
         }
         composable(Destinations.SETTINGS.route) {
+            val settingsViewModel = hiltViewModel<SettingsViewModel>()
             val state = settingsViewModel
                 .state
                 .collectAsState(initial = SettingsScreenViewState())
